CE-129 refactored code for sidenav
