+++ modules/examples/components/CodeBlock.component.tsx
