+++ modules/examples/components/CodeDescription.component.tsx
