+++ pages/_app.tsx
@@ -1,8 +1,8 @@
 import type { AppProps } from 'next/app';
 import { useState } from 'react';
 import 'tailwindcss/tailwind.css';
-import Header from '../components/Header';
-import SideNav from '../components/SideNav';
+import Header from '../common/components/Header';
+import SideNav from '../common/components/SideNav/SideNav.component';
 import exampleStepsContext from '../modules/examples/hooks/exampleSteps.context.hook';
 import '../styles/globals.css';
