+++ pages/index.tsx
@@ -1,17 +1,16 @@
 import type { NextPage } from 'next';
 import Head from 'next/head';
-import styles from '../styles/Home.module.css';
 
 const Home: NextPage = () => {
   return (
-    <div className={styles.container}>
+    <div>
       <Head>
         <title>Create Next App</title>
         <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
 
-      {/* <h1 className="text-4xl text-pink-500">Hello World in the Pink!</h1> */}
+      <h1 className="text-4xl text-pink-500">Hello World in the Pink!</h1>
     </div>
   );
 };
