+++ tailwind.config.js
@@ -4,7 +4,7 @@
 const defaultTheme = require('tailwindcss/defaultTheme');
 
 module.exports = {
-  content: ['pages/**/*.{ts,tsx}', 'components/**/*.{ts,tsx}'],
+  content: ['pages/**/*.{ts,tsx}', 'common/**/*.{ts,tsx}', 'modules/**/*.{ts,tsx}'],
   theme: {
     extend: {
       fontFamily: { sans: ['"Montserrat"', ...defaultTheme.fontFamily.sans] },
